requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
arrow==1.2.3
async-timeout==4.0.2
attrs==22.1.0
binaryornot==0.4.4
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cookiecutter==1.7.3
fastapi==0.79.1
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
loguru==0.6.0
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.2
nb-cli==0.6.7
nonebot-adapter-onebot==2.1.3
nonebot2==2.0.0b5
Pillow==9.2.0
poyo==0.5.0
prompt-toolkit==3.0.31
pydantic==1.9.2
pyfiglet==0.8.post1
pygtrie==2.5.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
text-unidecode==1.3
tomlkit==0.10.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.16.1
wcwidth==0.2.5
websockets==10.3
wget==3.2
win32-setctime==1.1.0
wincertstore==0.2
yarl==1.8.1
